corpus: fix 23 ruff findings from an unpinned ruff>=0.5, then pin it exactly
The numpy fix unblocked test collection, and CI's corpus job then failed at
`ruff check .` with 23 findings that showed up nowhere in local testing this
session. Root cause: pyproject.toml's dev extra pinned ruff>=0.5 (a floor, no
ceiling); this machine had ruff 0.15.11 installed, a fresh `pip install -e
"corpus[dev]"` on a GitHub runner resolved 0.16.6, which enables more rules by
default (UP037, RUF046, RUF059, SIM113, B017, BLE001, I001). Reproduced exactly
by upgrading locally to 0.16.6 before fixing anything.

Fixed all 23:
- ruff check --fix handled 20 mechanically (unnecessary quoted forward-refs now
  that `from __future__ import annotations` makes them unneeded, Callable moved
  to collections.abc, one unsorted import block).
- fingerprint.py's blind `except Exception` narrowed to the three real failure
  modes calling git can raise (OSError, CalledProcessError, TimeoutExpired).
- Two dead int(round(...)) casts in synth/generator.py - round() with no
  ndigits already returns int.
- Two unused unpacked variables in the vendored _sherpa_whisper_export.py
  (`qk` from qkv_attention, genuinely unused by both callers) renamed to `_qk`;
  a manual counter loop converted to enumerate().
- Two pytest.raises(Exception) widened to what each call site actually raises
  (ValueError for a checksum mismatch, EvalFoldSealed for the eval-fold gate) -
  checked the real exception before narrowing, not guessed.

Pinned ruff==0.16.6 exactly, replacing the open-ended floor, so a future ruff
release enabling new default rules becomes a deliberate version bump with its
own findings fixed in the same change, not a surprise on the next clean-runner
CI run.

Verified: ruff check corpus clean at 0.16.6; pytest corpus -q unchanged (70
passed, 1 skipped); repeated the from-scratch venv simulation used for the
numpy fix - both ruff check and pytest clean in complete isolation.

+++ corpus/src/corpus/synth/generator.py
@@ -51,10 +51,10 @@ def generate_utterance(
 
     sr = voice.sample_rate
     total_s = placed[-1].start_s + placed[-1].dur_s if placed else 0.0
-    buf = np.zeros(int(round(total_s * sr)), dtype=np.float64)
+    buf = np.zeros(round(total_s * sr), dtype=np.float64)
     for unit in placed:
         clip = voice.synth(unit.text)
-        start = int(round(unit.start_s * sr))
+        start = round(unit.start_s * sr)
         buf[start : start + len(clip)] = clip[: len(buf) - start]
 
     audio = apply_channel(buf, channel) if (channel.gain != 1.0 or channel.tilt or channel.noise_rms) else buf
